requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==41.0.0
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.3.0
faiss-cpu==1.7.4
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
Flask==2.3.2
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gradio==3.32.0
gradio_client==0.2.5
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.184
linkify-it-py==2.0.2
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
orjson==3.8.14
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pypdf==3.9.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.28.2
s3transfer==0.6.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
toolz==0.12.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
validators==0.20.0
waitress==2.1.2
weaviate-client==3.19.2
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
